add same logic to the deletePost method, where only the user who created can delete the post

+++ backend/controllers/feed.js
@@ -172,6 +172,11 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404
                 throw error
             }
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Auauthorized user.')
+                error.statusCode = 403
+                throw error
+            }
             deleteImageFile(post.imageUrl)
             return Post.findByIdAndRemove(postId)
         })
